Refactors state representation to 4D and stabilizes PPO training with squashing corrections
- Increases state dimension to 4 (Biomass, Nitrate, Product, and normalized Time) across the environment and agents.
- Implements Tanh Jacobian correction for actor log-probabilities and updates the memory buffer to store raw, unsquashed actions for stable evaluation.
- Transitions environment state calculations to float64 precision and synchronizes physical scaling constants across all modules.
- Refines the SPRL agent logic by introducing a mapping penalty to align latent intent with the safeguard's projected actions.
- Updates pretraining hyperparameters and loss weighting to improve identity mapping for safe trajectories.

+++ env.py
@@ -80,8 +80,8 @@ class PhycocyaninEnv:
         self.time_step_count = 0
         self.violation_count = 0
         
-        # Initial State: [1.0 g/L, 150 mg/L, 0.0 mg/L] [cite: 666]
-        self.state = np.array([1.0, 150.0, 0.0], dtype=np.float32)
+        # Initial State: [1.0 g/L, 150 mg/L, 0.0 mg/L]
+        self.state = np.array([1.0, 150.0, 0.0], dtype=np.float64)
         self.prev_action = np.zeros(2)
         
         # Metric Tracking
@@ -94,7 +94,7 @@ class PhycocyaninEnv:
         return self.get_state_norm()
 
     def get_state_norm(self):
-        norm_obs = np.zeros(4, dtype=np.float32)
+        norm_obs = np.zeros(4, dtype=np.float64)
         norm_obs[0] = self.state[0] / 6.0     
         norm_obs[1] = self.state[1] / 800.0   
         norm_obs[2] = self.state[2] / 0.2
